Move NewMonitorRequest from mapper to client.
The function NewMonitorRequest was a member of the mapper type, but it
doesn't depend on mapper object any more. It is not closely related to
mapper, so move it to the client module, where it is used.
